feat: fixing classes
