+++ TeachMeSkills.Group1/TeachMeSkills.Simulator.Cons/Program.cs
@@ -10,32 +10,9 @@ namespace TeachMeSkills.Simulator.Cons
 	{
 		public static void Main(string[] args)
 		{
-
-			var customer = new Customer();
-
 			var product = new Product();
 
-			Console.WriteLine(string.Format("Customer with {0} products has been created.", customer.Basket.Count));
-
 			Console.ReadLine();
-
-		}
-		public class Customer
-		{
-			private const int MaxProductsInBasket = 10;
-
-			public readonly List<Product> Basket = new List<Product>();
-
-			public Customer()
-			{
-				var random = new Random();
-				int productsToCreate = random.Next(1, 10);
-
-				for (int i = 0; i < productsToCreate; i++)
-				{
-					this.Basket.Add(new Product());
-				}
-			}
 		}
 		public class Product
 		{
