+++ TeachMeSkills.Group1/TeachMeSkills.Simulator.Core/Product.cs
@@ -16,7 +16,5 @@ namespace TeachMeSkills.Simulator.Core
 
         public int Count { get; set; }
 
-        public virtual ICollection<Sell> Sells { get; set; }
-
     }
 }
